w3c-reports/12092014/mac/runned-tests/html/syntax/parsing/html5lib_scripted_webkit01-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

html5lib Parser Test


FAIL html5lib_scripted_webkit01.html 3ff6ec1125852c7933bf6d89ecb375354e6e1b40 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_scripted_webkit01.html 46ae362de712eb9c55916de93110299dbbcb5726 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
